Move useEffect to currentForcast
